Move keep-related functions to separate file in local package
This is similar to the changes made in 7790204e4510336f8bed48b852c6c38f5e74ea72
for ethereum and celo packages, applied to the local package. The
existing bonded_ecdsa_keep.go file is repurposed to hold all keep-
related functions in preparation for moving to a BondedECDSAKeepHandle.
