+++ pkg/chain/local/local.go
@@ -12,13 +12,11 @@ import (
 	"time"
 
 	"github.com/keep-network/keep-core/pkg/chain/local"
-	"github.com/keep-network/keep-ecdsa/pkg/utils/byteutils"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/keep-network/keep-common/pkg/subscription"
 	corechain "github.com/keep-network/keep-core/pkg/chain"
 	"github.com/keep-network/keep-ecdsa/pkg/chain"
-	"github.com/keep-network/keep-ecdsa/pkg/ecdsa"
 
 	commonLocal "github.com/keep-network/keep-common/pkg/chain/local"
 )
@@ -161,160 +159,6 @@ func (lc *localChain) OnBondedECDSAKeepCreated(
 	})
 }
 
-// OnSignatureRequested is a callback that is invoked on-chain
-// when a keep's signature is requested.
-func (lc *localChain) OnSignatureRequested(
-	keepAddress common.Address,
-	handler func(event *chain.SignatureRequestedEvent),
-) (subscription.EventSubscription, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	handlerID := generateHandlerID()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return nil, fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	keep.signatureRequestedHandlers[handlerID] = handler
-
-	return subscription.NewEventSubscription(func() {
-		lc.localChainMutex.Lock()
-		defer lc.localChainMutex.Unlock()
-
-		delete(keep.signatureRequestedHandlers, handlerID)
-	}), nil
-}
-
-// SubmitKeepPublicKey checks if public key has been already submitted for given
-// keep address, if not it stores the key in a map.
-func (lc *localChain) SubmitKeepPublicKey(
-	keepAddress common.Address,
-	publicKey [64]byte,
-) error {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	if keep.publicKey != [64]byte{} {
-		return fmt.Errorf(
-			"public key already submitted for keep [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	keep.publicKey = publicKey
-
-	return nil
-}
-
-// SubmitSignature submits a signature to a keep contract deployed under a
-// given address.
-func (lc *localChain) SubmitSignature(
-	keepAddress common.Address,
-	signature *ecdsa.Signature,
-) error {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	// force the right workflow sequence
-	if keep.latestDigest == [32]byte{} {
-		return fmt.Errorf(
-			"keep [%s] is not awaiting for a signature",
-			keepAddress.String(),
-		)
-	}
-
-	rBytes, err := byteutils.BytesTo32Byte(signature.R.Bytes())
-	if err != nil {
-		return err
-	}
-
-	sBytes, err := byteutils.BytesTo32Byte(signature.S.Bytes())
-	if err != nil {
-		return err
-	}
-
-	keep.signatureSubmittedEvents = append(
-		keep.signatureSubmittedEvents,
-		&chain.SignatureSubmittedEvent{
-			Digest:     keep.latestDigest,
-			R:          rBytes,
-			S:          sBytes,
-			RecoveryID: uint8(signature.RecoveryID),
-		},
-	)
-
-	return nil
-}
-
-// IsAwaitingSignature checks if the keep is waiting for a signature to be
-// calculated for the given digest.
-func (lc *localChain) IsAwaitingSignature(
-	keepAddress common.Address,
-	digest [32]byte,
-) (bool, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return false, fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	return keep.latestDigest != [32]byte{}, nil
-}
-
-func (lc *localChain) GetPublicKey(keepAddress common.Address) ([]uint8, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return nil, fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	return keep.publicKey[:], nil
-}
-
-// IsActive checks for current state of a keep on-chain.
-func (lc *localChain) IsActive(keepAddress common.Address) (bool, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return false, fmt.Errorf("no keep with address [%v]", keepAddress)
-	}
-
-	return keep.status == active, nil
-}
-
 func (lc *localChain) BlockCounter() corechain.BlockCounter {
 	return lc.blockCounter
 }
@@ -347,138 +191,6 @@ func (lc *localChain) GetKeepCount() (*big.Int, error) {
 	defer lc.localChainMutex.Unlock()
 
 	return big.NewInt(int64(len(lc.keeps))), nil
-}
-
-func (lc *localChain) GetKeepAtIndex(
-	keepIndex *big.Int,
-) (common.Address, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	index := int(keepIndex.Uint64())
-
-	if index > len(lc.keepAddresses) {
-		return common.HexToAddress("0x0"), fmt.Errorf("out of bounds")
-	}
-
-	return lc.keepAddresses[index], nil
-}
-
-func (lc *localChain) OnKeepClosed(
-	keepAddress common.Address,
-	handler func(event *chain.KeepClosedEvent),
-) (subscription.EventSubscription, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	handlerID := generateHandlerID()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return nil, fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	keep.keepClosedHandlers[handlerID] = handler
-
-	return subscription.NewEventSubscription(func() {
-		lc.localChainMutex.Lock()
-		defer lc.localChainMutex.Unlock()
-
-		delete(keep.keepClosedHandlers, handlerID)
-	}), nil
-}
-
-func (lc *localChain) OnKeepTerminated(
-	keepAddress common.Address,
-	handler func(event *chain.KeepTerminatedEvent),
-) (subscription.EventSubscription, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	handlerID := generateHandlerID()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return nil, fmt.Errorf(
-			"failed to find keep with address: [%s]",
-			keepAddress.String(),
-		)
-	}
-
-	keep.keepTerminatedHandlers[handlerID] = handler
-
-	return subscription.NewEventSubscription(func() {
-		lc.localChainMutex.Lock()
-		defer lc.localChainMutex.Unlock()
-
-		delete(keep.keepTerminatedHandlers, handlerID)
-	}), nil
-}
-
-func (lc *localChain) OnConflictingPublicKeySubmitted(
-	keepAddress common.Address,
-	handler func(event *chain.ConflictingPublicKeySubmittedEvent),
-) (subscription.EventSubscription, error) {
-	panic("implement")
-}
-
-func (lc *localChain) OnPublicKeyPublished(
-	keepAddress common.Address,
-	handler func(event *chain.PublicKeyPublishedEvent),
-) (subscription.EventSubscription, error) {
-	panic("implement")
-}
-
-func (lc *localChain) LatestDigest(keepAddress common.Address) ([32]byte, error) {
-	panic("implement")
-}
-
-func (lc *localChain) SignatureRequestedBlock(
-	keepAddress common.Address,
-	digest [32]byte,
-) (uint64, error) {
-	panic("implement")
-}
-
-func (lc *localChain) GetMembers(
-	keepAddress common.Address,
-) ([]common.Address, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[keepAddress]
-	if !ok {
-		return nil, fmt.Errorf("no keep with address [%v]", keepAddress)
-	}
-	return keep.members, nil
-}
-
-func (lc *localChain) GetHonestThreshold(
-	keepAddress common.Address,
-) (uint64, error) {
-	panic("implement")
-}
-
-func (lc *localChain) GetOpenedTimestamp(keepAddress common.Address) (time.Time, error) {
-	panic("implement")
-}
-
-func (lc *localChain) PastSignatureSubmittedEvents(
-	keepAddress string,
-	startBlock uint64,
-) ([]*chain.SignatureSubmittedEvent, error) {
-	lc.localChainMutex.Lock()
-	defer lc.localChainMutex.Unlock()
-
-	keep, ok := lc.keeps[common.HexToAddress(keepAddress)]
-	if !ok {
-		return nil, fmt.Errorf("no keep with address [%v]", keepAddress)
-	}
-
-	return keep.signatureSubmittedEvents, nil
 }
 
 func (lc *localChain) BlockTimestamp(blockNumber *big.Int) (uint64, error) {
@@ -515,4 +227,68 @@ func generateAddress() common.Address {
 	// always `nil`.
 	rand.Read(address[:])
 	return address
+}
+
+func (lc *localChain) closeKeep(keepAddress common.Address) error {
+	lc.localChainMutex.Lock()
+	defer lc.localChainMutex.Unlock()
+
+	keep, ok := lc.keeps[keepAddress]
+	if !ok {
+		return fmt.Errorf(
+			"failed to find keep with address: [%s]",
+			keepAddress.String(),
+		)
+	}
+
+	if keep.status != active {
+		return fmt.Errorf("only active keeps can be closed")
+	}
+
+	keep.status = closed
+
+	keepClosedEvent := &chain.KeepClosedEvent{}
+
+	for _, handler := range keep.keepClosedHandlers {
+		go func(
+			handler func(event *chain.KeepClosedEvent),
+			keepClosedEvent *chain.KeepClosedEvent,
+		) {
+			handler(keepClosedEvent)
+		}(handler, keepClosedEvent)
+	}
+
+	return nil
+}
+
+func (lc *localChain) terminateKeep(keepAddress common.Address) error {
+	lc.localChainMutex.Lock()
+	defer lc.localChainMutex.Unlock()
+
+	keep, ok := lc.keeps[keepAddress]
+	if !ok {
+		return fmt.Errorf(
+			"failed to find keep with address: [%s]",
+			keepAddress.String(),
+		)
+	}
+
+	if keep.status != active {
+		return fmt.Errorf("only active keeps can be terminated")
+	}
+
+	keep.status = terminated
+
+	keepTerminatedEvent := &chain.KeepTerminatedEvent{}
+
+	for _, handler := range keep.keepTerminatedHandlers {
+		go func(
+			handler func(event *chain.KeepTerminatedEvent),
+			keepTerminatedEvent *chain.KeepTerminatedEvent,
+		) {
+			handler(keepTerminatedEvent)
+		}(handler, keepTerminatedEvent)
+	}
+
+	return nil
 }
